+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 ### Added
 
 - Added `JSONRPCVersion` constant
+- Added `ResponseSet` and `UnmarshalResponseSet()`
 
 ### Changed